nv_context helper struct, struct nv_view separates tui properties from buffer, nv_visual_row and nv_buff_line for more efficient line insertion, editing, etc. (WONT COMPILE)

+++ status.h
@@ -2,7 +2,7 @@
 #define NV_STATUS_H
 
 struct nv_status {
-    size_t height;
+    int height;
     char* format;
     char* string;
 };
